requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-socketio==5.13.0
sqlalchemy==2.0.43
psycopg[binary]==3.3.3
alembic==1.16.4
pydantic==2.11.7
pydantic-settings==2.10.1
email-validator==2.2.0
python-jose[cryptography]==3.5.0
passlib==1.7.4
python-multipart==0.0.20
redis==6.4.0
langchain==0.3.27
langchain-community==0.3.29
langchain-google-genai==2.1.10
faiss-cpu==1.12.0
httpx==0.28.1
tenacity==9.1.2
orjson==3.11.3
pytest==8.4.1
pytest-asyncio==1.1.0
